requirements.txt
fastapi
pydantic
uvicorn
requests
google-cloud-texttospeech
pillow
pydub
